+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning v2.0.0](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 31-Dec-2025
+
+### Changed
+
+- Fix formulas for backward difference form
+- Add formulas for central difference form
+- Minor metadata update
+- Add Mathematica programs for backward difference form
+- Add Mathematica programs for central difference form
+- Add Mathematica programs for Stirling form
+- Add Mathematica programs for Eulerian form
+- Update Mathematica documentation
+- Add Mathematica programs for multifold sums of powers binomial form
+- Add Mathematica programs for multifold sums of powers binomial form re-indexed
+
 ## [1.2.1] - 25-Dec-2025
 
 ### Changed